Feat(#22): 코드 수정
