+++ src/components/Tooltip/Tooltip.styled.ts
@@ -4,18 +4,18 @@ const Text = styled.span.attrs({
   className:
     'absolute bg-default-black text-default-white text-xs rounded py-1 px-2 opacity-0',
 })`
-  width: max-content;
-  bottom: 100%;
-  left: 50%;
-  transform: translateX(-50%);
-  margin-bottom: 5px;
-  transition: opacity 0.3s;
-  pointer-events: none;
+    width: max-content;
+    bottom: 100%;
+    left: 50%;
+    transform: translateX(-50%);
+    margin-bottom: 5px;
+    transition: opacity 0.3s;
+    pointer-events: none;
 
-  &:hover {
-    opacity: 1;
-    pointer-events: auto;
-  }
+    &:hover {
+        opacity: 1;
+        pointer-events: auto;
+    }
 `;
 
 const Container = styled.span.attrs({
