+++ src/components/group/invited/InvitedGroupInfo.tsx
@@ -11,10 +11,10 @@ import {
 } from '@/api/group/deleteGroupInvite';
 import patchGroupInviteAccepted from '@/api/group/patchGroupInviteAccepted';
 import Check from '@/assets/icons/check.svg';
-import X                   from '@/assets/icons/x.svg';
+import X from '@/assets/icons/x.svg';
+import ColorCircle from '@/components/common/ColorCircle';
+import Tooltip from '@/components/Tooltip/Tooltip';
 import { LabelColorsType } from '@/styles/colorThemes.ts';
-import ColorCircle         from '@/components/common/ColorCircle';
-import Tooltip from '@/components/tooltip/Tooltip';
 
 type InvitedGroupInfoProps = {
   group: Group;
