+++ src/mocks/calendar/v1/calendars/getCalendars.ts
@@ -39,6 +39,26 @@ const response: BaseResponse<CalendarsResponse> = {
           title: 'study group',
         },
       },
+      {
+        schedule_id: 3,
+        start_date: createDate(new Date(), 0, 0),
+        end_date: createDate(new Date(), 0, 1),
+        title: 'MSW: CS 스터디',
+        description: 'CS 스터디',
+        category: 'MEMBER',
+        label: null,
+        group: null,
+      },
+      {
+        schedule_id: 4,
+        start_date: createDate(new Date(), 0, 0),
+        end_date: createDate(new Date(), 0, 1),
+        title: 'MSW: 자기소개서 스터디',
+        description: 'CS 스터디',
+        category: 'MEMBER',
+        label: null,
+        group: null,
+      },
     ],
   },
 };
